Moved new path above the home to avoid path conflict
