+++ src/Components/DogDetail.js
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react';
-import { useParams } from 'react-router-dom/cjs/react-router-dom.min';
+import { useParams } from 'react-router-dom';
 import { fetchDogById } from '../services/dogs';
 
 export default function DogDetail() {
